third_party/WebKit/LayoutTests/svg/parser/whitespace-integer-expected.txt

CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47em".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47ex".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47px".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47in".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47cm".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47mm".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47pt".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47pc".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47%".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "0em".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "0ex".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "0px".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "0in".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "0cm".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "0mm".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "0pt".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "0pc".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "0%".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32em".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32ex".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32px".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32in".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32cm".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32mm".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32pt".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32pc".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32%".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245em".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245ex".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245px".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245in".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245cm".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245mm".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245pt".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245pc".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245%".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47em ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47ex ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47px ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47in ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47cm ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47mm ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47pt ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47pc ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47% ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "0em ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "0ex ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "0px ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "0in ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "0cm ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "0mm ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "0pt ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "0pc ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "0% ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32em ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32ex ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32px ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32in ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32cm ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32mm ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32pt ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32pc ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32% ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245em ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245ex ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245px ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245in ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245cm ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245mm ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245pt ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245pc ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245% ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47em   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47ex   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47px   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47in   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47cm   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47mm   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47pt   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47pc   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47%   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "0em   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "0ex   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "0px   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "0in   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "0cm   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "0mm   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "0pt   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "0pc   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "0%   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32em   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32ex   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32px   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32in   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32cm   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32mm   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32pt   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32pc   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32%   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245em   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245ex   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245px   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245in   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245cm   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245mm   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245pt   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245pc   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245%   ".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47a".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "0a".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32a".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245a".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47e".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "0e".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32e".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245e".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47foo".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "0foo".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32foo".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245foo".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47)90".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "0)90".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32)90".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245)90".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47em".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47ex".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47px".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47in".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47cm".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47mm".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47pt".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47pc".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47%".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0em".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0ex".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0px".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0in".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0cm".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0mm".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0pt".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0pc".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0%".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32em".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32ex".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32px".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32in".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32cm".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32mm".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32pt".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32pc".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32%".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245em".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245ex".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245px".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245in".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245cm".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245mm".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245pt".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245pc".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245%".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47em ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47ex ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47px ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47in ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47cm ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47mm ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47pt ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47pc ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47% ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0em ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0ex ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0px ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0in ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0cm ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0mm ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0pt ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0pc ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0% ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32em ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32ex ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32px ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32in ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32cm ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32mm ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32pt ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32pc ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32% ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245em ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245ex ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245px ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245in ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245cm ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245mm ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245pt ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245pc ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245% ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47em   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47ex   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47px   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47in   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47cm   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47mm   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47pt   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47pc   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47%   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0em   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0ex   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0px   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0in   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0cm   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0mm   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0pt   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0pc   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0%   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32em   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32ex   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32px   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32in   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32cm   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32mm   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32pt   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32pc   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32%   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245em   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245ex   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245px   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245in   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245cm   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245mm   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245pt   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245pc   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245%   ".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47 em".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "0 em".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32 em".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245 em".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47 ex".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "0 ex".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32 ex".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245 ex".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47 px".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "0 px".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32 px".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245 px".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47 in".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "0 in".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32 in".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245 in".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47 cm".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "0 cm".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32 cm".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245 cm".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47 mm".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "0 mm".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32 mm".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245 mm".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47 pt".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "0 pt".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32 pt".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245 pt".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47 pc".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "0 pc".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32 pc".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245 pc".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47 %".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "0 %".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32 %".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245 %".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47a".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0a".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32a".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245a".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47e".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0e".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32e".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245e".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47foo".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0foo".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32foo".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245foo".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, " -47)90".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, " 0)90".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, " +32)90".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, " 1241245)90".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47em".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47ex".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47px".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47in".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47cm".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47mm".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47pt".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47pc".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47%".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0em".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0ex".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0px".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0in".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0cm".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0mm".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0pt".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0pc".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0%".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32em".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32ex".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32px".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32in".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32cm".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32mm".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32pt".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32pc".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32%".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245em".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245ex".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245px".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245in".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245cm".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245mm".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245pt".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245pc".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245%".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47em ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47ex ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47px ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47in ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47cm ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47mm ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47pt ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47pc ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47% ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0em ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0ex ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0px ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0in ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0cm ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0mm ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0pt ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0pc ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0% ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32em ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32ex ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32px ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32in ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32cm ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32mm ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32pt ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32pc ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32% ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245em ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245ex ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245px ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245in ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245cm ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245mm ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245pt ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245pc ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245% ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47em   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47ex   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47px   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47in   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47cm   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47mm   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47pt   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47pc   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47%   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0em   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0ex   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0px   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0in   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0cm   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0mm   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0pt   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0pc   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0%   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32em   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32ex   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32px   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32in   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32cm   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32mm   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32pt   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32pc   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32%   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245em   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245ex   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245px   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245in   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245cm   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245mm   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245pt   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245pc   ".
CONSOLE ERROR: line 24: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245%   ".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47   em".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "0   em".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32   em".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245   em".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47   ex".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "0   ex".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32   ex".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245   ex".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47   px".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "0   px".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32   px".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245   px".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47   in".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "0   in".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32   in".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245   in".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47   cm".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "0   cm".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32   cm".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245   cm".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47   mm".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "0   mm".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32   mm".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245   mm".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47   pt".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "0   pt".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32   pt".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245   pt".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47   pc".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "0   pc".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32   pc".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245   pc".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "-47   %".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "0   %".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "+32   %".
CONSOLE ERROR: line 45: Error: <feTurbulence> attribute numOctaves: Expected integer, "1241245   %".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47a".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0a".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32a".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245a".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47e".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0e".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32e".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245e".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47foo".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0foo".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32foo".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245foo".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "   -47)90".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "   0)90".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "   +32)90".
CONSOLE ERROR: line 63: Error: <feTurbulence> attribute numOctaves: Expected integer, "   1241245)90".
This is a testharness.js-based test.
PASS Test <integer> valid value: -47
FAIL Test <integer> valid value: -47em assert_equals: expected -47 but got 0
FAIL Test <integer> valid value: -47ex assert_equals: expected -47 but got 0
FAIL Test <integer> valid value: -47px assert_equals: expected -47 but got 0
FAIL Test <integer> valid value: -47in assert_equals: expected -47 but got 0
FAIL Test <integer> valid value: -47cm assert_equals: expected -47 but got 0
FAIL Test <integer> valid value: -47mm assert_equals: expected -47 but got 0
FAIL Test <integer> valid value: -47pt assert_equals: expected -47 but got 0
FAIL Test <integer> valid value: -47pc assert_equals: expected -47 but got 0
FAIL Test <integer> valid value: -47% assert_equals: expected -47 but got 0
PASS Test <integer> valid value: 0
PASS Test <integer> valid value: 0em
PASS Test <integer> valid value: 0ex
PASS Test <integer> valid value: 0px
PASS Test <integer> valid value: 0in
PASS Test <integer> valid value: 0cm
PASS Test <integer> valid value: 0mm
PASS Test <integer> valid value: 0pt
PASS Test <integer> valid value: 0pc
PASS Test <integer> valid value: 0%
PASS Test <integer> valid value: +32
FAIL Test <integer> valid value: +32em assert_equals: expected 32 but got 0
FAIL Test <integer> valid value: +32ex assert_equals: expected 32 but got 0
FAIL Test <integer> valid value: +32px assert_equals: expected 32 but got 0
FAIL Test <integer> valid value: +32in assert_equals: expected 32 but got 0
FAIL Test <integer> valid value: +32cm assert_equals: expected 32 but got 0
FAIL Test <integer> valid value: +32mm assert_equals: expected 32 but got 0
FAIL Test <integer> valid value: +32pt assert_equals: expected 32 but got 0
FAIL Test <integer> valid value: +32pc assert_equals: expected 32 but got 0
FAIL Test <integer> valid value: +32% assert_equals: expected 32 but got 0
PASS Test <integer> valid value: 1241245
FAIL Test <integer> valid value: 1241245em assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value: 1241245ex assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value: 1241245px assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value: 1241245in assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value: 1241245cm assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value: 1241245mm assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value: 1241245pt assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value: 1241245pc assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value: 1241245% assert_equals: expected 1241245 but got 0
PASS Test <integer> valid value: -47 
FAIL Test <integer> valid value: -47em  assert_equals: expected -47 but got 0
FAIL Test <integer> valid value: -47ex  assert_equals: expected -47 but got 0
FAIL Test <integer> valid value: -47px  assert_equals: expected -47 but got 0
FAIL Test <integer> valid value: -47in  assert_equals: expected -47 but got 0
FAIL Test <integer> valid value: -47cm  assert_equals: expected -47 but got 0
FAIL Test <integer> valid value: -47mm  assert_equals: expected -47 but got 0
FAIL Test <integer> valid value: -47pt  assert_equals: expected -47 but got 0
FAIL Test <integer> valid value: -47pc  assert_equals: expected -47 but got 0
FAIL Test <integer> valid value: -47%  assert_equals: expected -47 but got 0
PASS Test <integer> valid value: 0 
PASS Test <integer> valid value: 0em 
PASS Test <integer> valid value: 0ex 
PASS Test <integer> valid value: 0px 
PASS Test <integer> valid value: 0in 
PASS Test <integer> valid value: 0cm 
PASS Test <integer> valid value: 0mm 
PASS Test <integer> valid value: 0pt 
PASS Test <integer> valid value: 0pc 
PASS Test <integer> valid value: 0% 
PASS Test <integer> valid value: +32 
FAIL Test <integer> valid value: +32em  assert_equals: expected 32 but got 0
FAIL Test <integer> valid value: +32ex  assert_equals: expected 32 but got 0
FAIL Test <integer> valid value: +32px  assert_equals: expected 32 but got 0
FAIL Test <integer> valid value: +32in  assert_equals: expected 32 but got 0
FAIL Test <integer> valid value: +32cm  assert_equals: expected 32 but got 0
FAIL Test <integer> valid value: +32mm  assert_equals: expected 32 but got 0
FAIL Test <integer> valid value: +32pt  assert_equals: expected 32 but got 0
FAIL Test <integer> valid value: +32pc  assert_equals: expected 32 but got 0
FAIL Test <integer> valid value: +32%  assert_equals: expected 32 but got 0
PASS Test <integer> valid value: 1241245 
FAIL Test <integer> valid value: 1241245em  assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value: 1241245ex  assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value: 1241245px  assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value: 1241245in  assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value: 1241245cm  assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value: 1241245mm  assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value: 1241245pt  assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value: 1241245pc  assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value: 1241245%  assert_equals: expected 1241245 but got 0
PASS Test <integer> valid value: -47   
FAIL Test <integer> valid value: -47em    assert_equals: expected -47 but got 0
FAIL Test <integer> valid value: -47ex    assert_equals: expected -47 but got 0
FAIL Test <integer> valid value: -47px    assert_equals: expected -47 but got 0
FAIL Test <integer> valid value: -47in    assert_equals: expected -47 but got 0
FAIL Test <integer> valid value: -47cm    assert_equals: expected -47 but got 0
FAIL Test <integer> valid value: -47mm    assert_equals: expected -47 but got 0
FAIL Test <integer> valid value: -47pt    assert_equals: expected -47 but got 0
FAIL Test <integer> valid value: -47pc    assert_equals: expected -47 but got 0
FAIL Test <integer> valid value: -47%    assert_equals: expected -47 but got 0
PASS Test <integer> valid value: 0   
PASS Test <integer> valid value: 0em   
PASS Test <integer> valid value: 0ex   
PASS Test <integer> valid value: 0px   
PASS Test <integer> valid value: 0in   
PASS Test <integer> valid value: 0cm   
PASS Test <integer> valid value: 0mm   
PASS Test <integer> valid value: 0pt   
PASS Test <integer> valid value: 0pc   
PASS Test <integer> valid value: 0%   
PASS Test <integer> valid value: +32   
FAIL Test <integer> valid value: +32em    assert_equals: expected 32 but got 0
FAIL Test <integer> valid value: +32ex    assert_equals: expected 32 but got 0
FAIL Test <integer> valid value: +32px    assert_equals: expected 32 but got 0
FAIL Test <integer> valid value: +32in    assert_equals: expected 32 but got 0
FAIL Test <integer> valid value: +32cm    assert_equals: expected 32 but got 0
FAIL Test <integer> valid value: +32mm    assert_equals: expected 32 but got 0
FAIL Test <integer> valid value: +32pt    assert_equals: expected 32 but got 0
FAIL Test <integer> valid value: +32pc    assert_equals: expected 32 but got 0
FAIL Test <integer> valid value: +32%    assert_equals: expected 32 but got 0
PASS Test <integer> valid value: 1241245   
FAIL Test <integer> valid value: 1241245em    assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value: 1241245ex    assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value: 1241245px    assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value: 1241245in    assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value: 1241245cm    assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value: 1241245mm    assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value: 1241245pt    assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value: 1241245pc    assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value: 1241245%    assert_equals: expected 1241245 but got 0
PASS Test <integer> trailing garbage, value: -47a
PASS Test <integer> trailing garbage, value: 0a
PASS Test <integer> trailing garbage, value: +32a
PASS Test <integer> trailing garbage, value: 1241245a
PASS Test <integer> trailing garbage, value: -47e
PASS Test <integer> trailing garbage, value: 0e
PASS Test <integer> trailing garbage, value: +32e
PASS Test <integer> trailing garbage, value: 1241245e
PASS Test <integer> trailing garbage, value: -47foo
PASS Test <integer> trailing garbage, value: 0foo
PASS Test <integer> trailing garbage, value: +32foo
PASS Test <integer> trailing garbage, value: 1241245foo
PASS Test <integer> trailing garbage, value: -47)90
PASS Test <integer> trailing garbage, value: 0)90
PASS Test <integer> trailing garbage, value: +32)90
PASS Test <integer> trailing garbage, value: 1241245)90
PASS Test <integer> valid value:  -47
FAIL Test <integer> valid value:  -47em assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:  -47ex assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:  -47px assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:  -47in assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:  -47cm assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:  -47mm assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:  -47pt assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:  -47pc assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:  -47% assert_equals: expected -47 but got 0
PASS Test <integer> valid value:  0
PASS Test <integer> valid value:  0em
PASS Test <integer> valid value:  0ex
PASS Test <integer> valid value:  0px
PASS Test <integer> valid value:  0in
PASS Test <integer> valid value:  0cm
PASS Test <integer> valid value:  0mm
PASS Test <integer> valid value:  0pt
PASS Test <integer> valid value:  0pc
PASS Test <integer> valid value:  0%
PASS Test <integer> valid value:  +32
FAIL Test <integer> valid value:  +32em assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:  +32ex assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:  +32px assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:  +32in assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:  +32cm assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:  +32mm assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:  +32pt assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:  +32pc assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:  +32% assert_equals: expected 32 but got 0
PASS Test <integer> valid value:  1241245
FAIL Test <integer> valid value:  1241245em assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:  1241245ex assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:  1241245px assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:  1241245in assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:  1241245cm assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:  1241245mm assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:  1241245pt assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:  1241245pc assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:  1241245% assert_equals: expected 1241245 but got 0
PASS Test <integer> valid value:  -47 
FAIL Test <integer> valid value:  -47em  assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:  -47ex  assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:  -47px  assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:  -47in  assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:  -47cm  assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:  -47mm  assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:  -47pt  assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:  -47pc  assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:  -47%  assert_equals: expected -47 but got 0
PASS Test <integer> valid value:  0 
PASS Test <integer> valid value:  0em 
PASS Test <integer> valid value:  0ex 
PASS Test <integer> valid value:  0px 
PASS Test <integer> valid value:  0in 
PASS Test <integer> valid value:  0cm 
PASS Test <integer> valid value:  0mm 
PASS Test <integer> valid value:  0pt 
PASS Test <integer> valid value:  0pc 
PASS Test <integer> valid value:  0% 
PASS Test <integer> valid value:  +32 
FAIL Test <integer> valid value:  +32em  assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:  +32ex  assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:  +32px  assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:  +32in  assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:  +32cm  assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:  +32mm  assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:  +32pt  assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:  +32pc  assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:  +32%  assert_equals: expected 32 but got 0
PASS Test <integer> valid value:  1241245 
FAIL Test <integer> valid value:  1241245em  assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:  1241245ex  assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:  1241245px  assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:  1241245in  assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:  1241245cm  assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:  1241245mm  assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:  1241245pt  assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:  1241245pc  assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:  1241245%  assert_equals: expected 1241245 but got 0
PASS Test <integer> valid value:  -47   
FAIL Test <integer> valid value:  -47em    assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:  -47ex    assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:  -47px    assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:  -47in    assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:  -47cm    assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:  -47mm    assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:  -47pt    assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:  -47pc    assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:  -47%    assert_equals: expected -47 but got 0
PASS Test <integer> valid value:  0   
PASS Test <integer> valid value:  0em   
PASS Test <integer> valid value:  0ex   
PASS Test <integer> valid value:  0px   
PASS Test <integer> valid value:  0in   
PASS Test <integer> valid value:  0cm   
PASS Test <integer> valid value:  0mm   
PASS Test <integer> valid value:  0pt   
PASS Test <integer> valid value:  0pc   
PASS Test <integer> valid value:  0%   
PASS Test <integer> valid value:  +32   
FAIL Test <integer> valid value:  +32em    assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:  +32ex    assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:  +32px    assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:  +32in    assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:  +32cm    assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:  +32mm    assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:  +32pt    assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:  +32pc    assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:  +32%    assert_equals: expected 32 but got 0
PASS Test <integer> valid value:  1241245   
FAIL Test <integer> valid value:  1241245em    assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:  1241245ex    assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:  1241245px    assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:  1241245in    assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:  1241245cm    assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:  1241245mm    assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:  1241245pt    assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:  1241245pc    assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:  1241245%    assert_equals: expected 1241245 but got 0
PASS Test <integer> WS invalid value: -47 em
PASS Test <integer> WS invalid value: 0 em
PASS Test <integer> WS invalid value: +32 em
PASS Test <integer> WS invalid value: 1241245 em
PASS Test <integer> WS invalid value: -47 ex
PASS Test <integer> WS invalid value: 0 ex
PASS Test <integer> WS invalid value: +32 ex
PASS Test <integer> WS invalid value: 1241245 ex
PASS Test <integer> WS invalid value: -47 px
PASS Test <integer> WS invalid value: 0 px
PASS Test <integer> WS invalid value: +32 px
PASS Test <integer> WS invalid value: 1241245 px
PASS Test <integer> WS invalid value: -47 in
PASS Test <integer> WS invalid value: 0 in
PASS Test <integer> WS invalid value: +32 in
PASS Test <integer> WS invalid value: 1241245 in
PASS Test <integer> WS invalid value: -47 cm
PASS Test <integer> WS invalid value: 0 cm
PASS Test <integer> WS invalid value: +32 cm
PASS Test <integer> WS invalid value: 1241245 cm
PASS Test <integer> WS invalid value: -47 mm
PASS Test <integer> WS invalid value: 0 mm
PASS Test <integer> WS invalid value: +32 mm
PASS Test <integer> WS invalid value: 1241245 mm
PASS Test <integer> WS invalid value: -47 pt
PASS Test <integer> WS invalid value: 0 pt
PASS Test <integer> WS invalid value: +32 pt
PASS Test <integer> WS invalid value: 1241245 pt
PASS Test <integer> WS invalid value: -47 pc
PASS Test <integer> WS invalid value: 0 pc
PASS Test <integer> WS invalid value: +32 pc
PASS Test <integer> WS invalid value: 1241245 pc
PASS Test <integer> WS invalid value: -47 %
PASS Test <integer> WS invalid value: 0 %
PASS Test <integer> WS invalid value: +32 %
PASS Test <integer> WS invalid value: 1241245 %
PASS Test <integer> trailing garbage, value:  -47a
PASS Test <integer> trailing garbage, value:  0a
PASS Test <integer> trailing garbage, value:  +32a
PASS Test <integer> trailing garbage, value:  1241245a
PASS Test <integer> trailing garbage, value:  -47e
PASS Test <integer> trailing garbage, value:  0e
PASS Test <integer> trailing garbage, value:  +32e
PASS Test <integer> trailing garbage, value:  1241245e
PASS Test <integer> trailing garbage, value:  -47foo
PASS Test <integer> trailing garbage, value:  0foo
PASS Test <integer> trailing garbage, value:  +32foo
PASS Test <integer> trailing garbage, value:  1241245foo
PASS Test <integer> trailing garbage, value:  -47)90
PASS Test <integer> trailing garbage, value:  0)90
PASS Test <integer> trailing garbage, value:  +32)90
PASS Test <integer> trailing garbage, value:  1241245)90
PASS Test <integer> valid value:    -47
FAIL Test <integer> valid value:    -47em assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:    -47ex assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:    -47px assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:    -47in assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:    -47cm assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:    -47mm assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:    -47pt assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:    -47pc assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:    -47% assert_equals: expected -47 but got 0
PASS Test <integer> valid value:    0
PASS Test <integer> valid value:    0em
PASS Test <integer> valid value:    0ex
PASS Test <integer> valid value:    0px
PASS Test <integer> valid value:    0in
PASS Test <integer> valid value:    0cm
PASS Test <integer> valid value:    0mm
PASS Test <integer> valid value:    0pt
PASS Test <integer> valid value:    0pc
PASS Test <integer> valid value:    0%
PASS Test <integer> valid value:    +32
FAIL Test <integer> valid value:    +32em assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:    +32ex assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:    +32px assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:    +32in assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:    +32cm assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:    +32mm assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:    +32pt assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:    +32pc assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:    +32% assert_equals: expected 32 but got 0
PASS Test <integer> valid value:    1241245
FAIL Test <integer> valid value:    1241245em assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:    1241245ex assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:    1241245px assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:    1241245in assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:    1241245cm assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:    1241245mm assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:    1241245pt assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:    1241245pc assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:    1241245% assert_equals: expected 1241245 but got 0
PASS Test <integer> valid value:    -47 
FAIL Test <integer> valid value:    -47em  assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:    -47ex  assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:    -47px  assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:    -47in  assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:    -47cm  assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:    -47mm  assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:    -47pt  assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:    -47pc  assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:    -47%  assert_equals: expected -47 but got 0
PASS Test <integer> valid value:    0 
PASS Test <integer> valid value:    0em 
PASS Test <integer> valid value:    0ex 
PASS Test <integer> valid value:    0px 
PASS Test <integer> valid value:    0in 
PASS Test <integer> valid value:    0cm 
PASS Test <integer> valid value:    0mm 
PASS Test <integer> valid value:    0pt 
PASS Test <integer> valid value:    0pc 
PASS Test <integer> valid value:    0% 
PASS Test <integer> valid value:    +32 
FAIL Test <integer> valid value:    +32em  assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:    +32ex  assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:    +32px  assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:    +32in  assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:    +32cm  assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:    +32mm  assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:    +32pt  assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:    +32pc  assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:    +32%  assert_equals: expected 32 but got 0
PASS Test <integer> valid value:    1241245 
FAIL Test <integer> valid value:    1241245em  assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:    1241245ex  assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:    1241245px  assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:    1241245in  assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:    1241245cm  assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:    1241245mm  assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:    1241245pt  assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:    1241245pc  assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:    1241245%  assert_equals: expected 1241245 but got 0
PASS Test <integer> valid value:    -47   
FAIL Test <integer> valid value:    -47em    assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:    -47ex    assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:    -47px    assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:    -47in    assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:    -47cm    assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:    -47mm    assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:    -47pt    assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:    -47pc    assert_equals: expected -47 but got 0
FAIL Test <integer> valid value:    -47%    assert_equals: expected -47 but got 0
PASS Test <integer> valid value:    0   
PASS Test <integer> valid value:    0em   
PASS Test <integer> valid value:    0ex   
PASS Test <integer> valid value:    0px   
PASS Test <integer> valid value:    0in   
PASS Test <integer> valid value:    0cm   
PASS Test <integer> valid value:    0mm   
PASS Test <integer> valid value:    0pt   
PASS Test <integer> valid value:    0pc   
PASS Test <integer> valid value:    0%   
PASS Test <integer> valid value:    +32   
FAIL Test <integer> valid value:    +32em    assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:    +32ex    assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:    +32px    assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:    +32in    assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:    +32cm    assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:    +32mm    assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:    +32pt    assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:    +32pc    assert_equals: expected 32 but got 0
FAIL Test <integer> valid value:    +32%    assert_equals: expected 32 but got 0
PASS Test <integer> valid value:    1241245   
FAIL Test <integer> valid value:    1241245em    assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:    1241245ex    assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:    1241245px    assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:    1241245in    assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:    1241245cm    assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:    1241245mm    assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:    1241245pt    assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:    1241245pc    assert_equals: expected 1241245 but got 0
FAIL Test <integer> valid value:    1241245%    assert_equals: expected 1241245 but got 0
PASS Test <integer> WS invalid value: -47   em
PASS Test <integer> WS invalid value: 0   em
PASS Test <integer> WS invalid value: +32   em
PASS Test <integer> WS invalid value: 1241245   em
PASS Test <integer> WS invalid value: -47   ex
PASS Test <integer> WS invalid value: 0   ex
PASS Test <integer> WS invalid value: +32   ex
PASS Test <integer> WS invalid value: 1241245   ex
PASS Test <integer> WS invalid value: -47   px
PASS Test <integer> WS invalid value: 0   px
PASS Test <integer> WS invalid value: +32   px
PASS Test <integer> WS invalid value: 1241245   px
PASS Test <integer> WS invalid value: -47   in
PASS Test <integer> WS invalid value: 0   in
PASS Test <integer> WS invalid value: +32   in
PASS Test <integer> WS invalid value: 1241245   in
PASS Test <integer> WS invalid value: -47   cm
PASS Test <integer> WS invalid value: 0   cm
PASS Test <integer> WS invalid value: +32   cm
PASS Test <integer> WS invalid value: 1241245   cm
PASS Test <integer> WS invalid value: -47   mm
PASS Test <integer> WS invalid value: 0   mm
PASS Test <integer> WS invalid value: +32   mm
PASS Test <integer> WS invalid value: 1241245   mm
PASS Test <integer> WS invalid value: -47   pt
PASS Test <integer> WS invalid value: 0   pt
PASS Test <integer> WS invalid value: +32   pt
PASS Test <integer> WS invalid value: 1241245   pt
PASS Test <integer> WS invalid value: -47   pc
PASS Test <integer> WS invalid value: 0   pc
PASS Test <integer> WS invalid value: +32   pc
PASS Test <integer> WS invalid value: 1241245   pc
PASS Test <integer> WS invalid value: -47   %
PASS Test <integer> WS invalid value: 0   %
PASS Test <integer> WS invalid value: +32   %
PASS Test <integer> WS invalid value: 1241245   %
PASS Test <integer> trailing garbage, value:    -47a
PASS Test <integer> trailing garbage, value:    0a
PASS Test <integer> trailing garbage, value:    +32a
PASS Test <integer> trailing garbage, value:    1241245a
PASS Test <integer> trailing garbage, value:    -47e
PASS Test <integer> trailing garbage, value:    0e
PASS Test <integer> trailing garbage, value:    +32e
PASS Test <integer> trailing garbage, value:    1241245e
PASS Test <integer> trailing garbage, value:    -47foo
PASS Test <integer> trailing garbage, value:    0foo
PASS Test <integer> trailing garbage, value:    +32foo
PASS Test <integer> trailing garbage, value:    1241245foo
PASS Test <integer> trailing garbage, value:    -47)90
PASS Test <integer> trailing garbage, value:    0)90
PASS Test <integer> trailing garbage, value:    +32)90
PASS Test <integer> trailing garbage, value:    1241245)90
Harness: the test ran to completion.
